refactor folder structure for server-side code
