+++ insta-ui/src/components/LLM/hooks/useCodeGeneration.ts
@@ -1,4 +1,3 @@
-// app/hooks/useCodeGeneration.ts
 'use client';
 
 import { useChat } from 'ai/react';
