+++ insta-ui/src/server/constants.ts
@@ -1,4 +1,3 @@
-// app/server/llm/constants.ts
 export const MAX_ATTEMPTS = 10;
 export const MAX_TOKENS = 1000;
 export const TEMPERATURE = 0.7;
